+++ src/Taxes/Drivers/TaxEngineManager.php
@@ -43,6 +43,11 @@ class TaxEngineManager
     ) {
     }
 
+    public static function getDrivers(): array
+    {
+        return self::$drivers;
+    }
+
     public function __call(string $method, array $arguments)
     {
         return $this->driver()->$method(...$arguments);