+++ Changelog.md
@@ -12,6 +12,7 @@
    - `BasePaymentEvent` class => it applies to all payment events
 - Added shipment event classes (created, delivered, ready, lost, etc)
 - Added support to media settings `background` and `format` (passing "original" calls `keepOriginalFormat()`)
+- Added the `TaxEngineManager::getDrivers()` static method
 - Added the registration of `promotion` entry to the Eloquent morph map
 - Added the cart item percent discount promotion action type
 - Changed the cart fixed and percent discount promotion action types to only act on carts, not on cart items